+++ CHANGELOG.md
@@ -56,6 +56,7 @@
     * All `setWebhook` extensions was marked as deprecated, renamed and replaced into `TelegramBotAPI-extensions-utils`
     * Typealias `ExceptionHandler` was added - it will be used for `handleSafely`
     * `SetWebhook` factories signatures was changed (backward compatibility was not broken)
+    * `executeUnsafe` now working differently
 * `TelegramBotAPI-extensions-api`:
     * Long Polling extensions now are deprecated in this project. It was replaced into `TelegramBotAPI-extensions-utils`
 * `TelegramBotAPI-extensions-utils`: